pyproject.toml
[tool.poetry]
name = "ils-middleware"
version = "0.3.0"
description = "Sinopia ILS Middleware using Apache Airflow"
authors = ["Jeremy Nelson <jpnelson@stanford.edu>", "Aaron Collier <aaron.collier@stanford.edu", "John Martin <suntzu@stanford.edu>"]
license = "Apache2"

[tool.poetry.dependencies]
python = "^3.10"
pymarc = "^4.1.1"
apache-airflow = "^2.6.1"
apache-airflow-providers-amazon = "^2.2.0"
SQLAlchemy = "^1.3.18"
typing-extensions = ">=4.0.0"
rdflib = "6.3.2"
honeybadger = "^0.7.1"
types-requests = "^2.26.1"
lxml = "^4.6.4"
folio-uuid = "^0.2.0"
folioclient = "^0.50.0"
numpy = "^1.23.3"
pendulum = "2.1.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^3.9.2"
mypy = "^1.3.0"
pytest = "^6.2"
pytest-cov = "^2.12"
pytest-mock = "^3.6.1"
types-mock = "^4.0.1"
requests-mock = "^1.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"